refactor: 워크스페이스 네브바 컴포넌트로 변경
