+++ src/app/(afterLogin)/workspace/layout.tsx
@@ -1,12 +1,13 @@
 'use client';
 
-import { ReactNode, useEffect, useState } from 'react';
+import { ReactNode } from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
 import { usePathname, useSelectedLayoutSegment } from 'next/navigation';
 
 import settings from '@/../public/svgs/workspace/settings.svg';
 import BackArrow from '../_components/BackArrow';
+import NavbarItems from './_components/NavbarItems';
 
 type Props = {
   children: ReactNode;
@@ -16,37 +17,9 @@ export default function Layout({ children }: Props) {
   const workspaceId = useSelectedLayoutSegment();
   const workspaceIdNumber = Number(workspaceId);
 
-  const [activeSegment, setActiveSegment] = useState<string | null>(null);
   const pathName = usePathname();
   const segments = pathName.split('/');
   const currentSegment = segments[segments.length - 1];
-
-  useEffect(() => {
-    setActiveSegment(currentSegment);
-  }, [currentSegment]);
-
-  const navItems = [
-    {
-      name: '그룹홈',
-      path: `/workspace/${workspaceIdNumber}`,
-      segments: [workspaceId, 'workspaceHistory'],
-    },
-    {
-      name: '운동하기',
-      path: `/workspace/${workspaceIdNumber}/workout`,
-      segments: ['workout', 'register'],
-    },
-    {
-      name: '그룹채팅',
-      path: `/workspace/${workspaceIdNumber}/chat`,
-      segments: 'chat',
-    },
-    {
-      name: '운동인증',
-      path: `/workspace/${workspaceIdNumber}/auth`,
-      segments: 'auth',
-    },
-  ];
 
   return (
     <div className='h-full'>
@@ -62,23 +35,10 @@ export default function Layout({ children }: Props) {
         <nav className='my-3'>
           <hr className='border-1 border-[#E5E7EB] w-screen -mx-4' />
           <ul className='flex text-sm gap-x-11 sm:gap-x-8 lg:gap-x-12 justify-center my-2.5 text-[#E5E7EB]'>
-            {navItems.map((navItem) => (
-              <Link href={navItem.path} key={navItem.name}>
-                <li
-                  className={`${
-                    navItem.segments &&
-                    activeSegment !== null &&
-                    navItem.segments.includes(activeSegment)
-                      ? 'text-[#4B5563]'
-                      : navItem.segments === activeSegment
-                      ? 'text-[#4B5563]'
-                      : 'text-[#E5E7EB]'
-                  }`}
-                >
-                  {navItem.name}
-                </li>
-              </Link>
-            ))}
+            <NavbarItems
+              workspaceId={workspaceIdNumber}
+              currentSegment={currentSegment}
+            />
           </ul>
           <hr className='border-1 border-[#E5E7EB] w-screen -mx-4' />
         </nav>
